DJANGO-NINJA/back-end/requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-cors-headers==4.3.0
django-ninja==0.22.2
django-storages==1.14.2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
kombu==5.3.2
prompt-toolkit==3.0.39
pydantic==1.10.13
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
